requirements.txt
fastapi
uvicorn
groq
pycountry
resend
phidata 
python-dotenv
duckduckgo-search